For Pulse, send ESP only of the same IP protocol as we're connected over
It really seems that when we're connected over Legacy IP, it only accepts
Legacy IP packets in ESP. And when we're connected over IPv6, it only
accepts IPv6 packets in ESP.

Unless we're supposed to (have a public IPv6 route to the server and)
establish ESP over *both* Legacy IP and IPv6, and send the right packets
over each? Which would be insane.

Someone please tell me this isn't true. But it matches the behaviour
of the Windows client too.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>